# api/docs/swagger.yaml
definitions:
...
  entity.CreateUserResponse:
    properties:
      guid:
        type: string
    type: object
  entity.CreateVendorUserResponse:
    properties:
      guid:
        type: string
    type: object
...
  entity.DeleteUserRequest:
    properties:
      guid:
        type: string
    type: object
  entity.DeleteVendorUserRequest:
    properties:
      guid:
        type: string
    type: object
...
  entity.GetOrderResponse:
    properties:
      address:
        type: string
      branch_address:
        type: string
      branch_id:
        type: string
      branch_name:
        type: string
      comission:
        type: number
      comment:
        type: string
      courier_id:
        type: string
      created_at:
        type: string
      delivery_price:
        type: number
      delivery_time:
        type: number
      delivery_type:
        type: string
      guid:
        type: string
      image:
        type: string
      merchant_name:
        type: string
      payment_type:
        type: string
      products:
        items:
          $ref: '#/definitions/entity.OrderProducts'
        type: array
      status:
        type: string
      user_id:
        type: string
    type: object
...
  entity.GetUserOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/entity.GetOrderResponse'
        type: array
    type: object
  entity.GetUserResponse:
    properties:
      created_at:
        type: string
      guid:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  entity.GetVendorUserResponse:
    properties:
      created_at:
        type: string
      guid:
        type: string
      login:
        type: string
      merchant_branch_id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  entity.ListUserResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/entity.GetUserResponse'
        type: array
    type: object
  entity.ListVendorUserResponse:
    properties:
      vendor_users:
        items:
          $ref: '#/definitions/entity.GetVendorUserResponse'
        type: array
    type: object
...
  entity.LoginUserRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  entity.LoginUserResponse:
    properties:
      exist:
        type: boolean
      guid:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  entity.LoginVendorUserRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  entity.LoginVendorUserResponse:
    properties:
      exist:
        type: boolean
      guid:
        type: string
      merchant_id:
        type: string
      name:
        type: string
    type: object
  entity.Option:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
...
  entity.OrderProducts:
    properties:
      category_id:
        type: string
      count:
        type: integer
      created_at:
        type: string
      guid:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/entity.Option'
        type: array
      price:
        type: number
    type: object
  entity.ResponseOK:
    properties:
      message:
        type: object
    type: object
...
  entity.UpdateUserRequest:
    properties:
      guid:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  entity.UpdateUserResponse:
    properties:
      guid:
        type: string
    type: object
  entity.UpdateVendorUserRequest:
    properties:
      guid:
        type: string
      login:
        type: string
      merchant_branch_id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  entity.UpdateVendorUserResponse:
    properties:
      guid:
        type: string
    type: object
  models.ResponseError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.User:
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.VendorUser:
    properties:
      login:
        type: string
      merchant_branch_id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/user/{guid}/orders:
    get:
      consumes:
      - application/json
      description: API that get user orders
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetUserOrdersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get user orders
      tags:
      - user
  /v1/user/create:
    post:
      consumes:
      - application/json
      description: API that create user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create user
      tags:
      - user
  /v1/user/delete:
    delete:
      consumes:
      - application/json
      description: API that delete user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete user
      tags:
      - user
  /v1/user/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get user
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get user
      tags:
      - user
  /v1/user/list:
    get:
      consumes:
      - application/json
      description: API list users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List users
      tags:
      - user
  /v1/user/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Login user
      tags:
      - user
  /v1/user/update:
    put:
      consumes:
      - application/json
      description: API that update user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update user
      tags:
      - user
  /v1/vendor-user/create:
    post:
      consumes:
      - application/json
      description: API that create vendor-user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.VendorUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CreateVendorUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create vendor-user
      tags:
      - vendor-user
  /v1/vendor-user/delete:
    delete:
      consumes:
      - application/json
      description: API that delete vendor-user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteVendorUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ResponseOK'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete vendor-user
      tags:
      - vendor-user
  /v1/vendor-user/get/{guid}:
    get:
      consumes:
      - application/json
      description: API that get vendor-user
      parameters:
      - description: guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetVendorUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get vendor-user
      tags:
      - vendor-user
  /v1/vendor-user/list:
    get:
      consumes:
      - application/json
      description: API list vendor-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ListVendorUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: List vendor-users
      tags:
      - vendor-user
  /v1/vendor-user/login:
    post:
      consumes:
      - application/json
      description: Login vendor user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.LoginVendorUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginVendorUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Login vendor user
      tags:
      - vendor-user
  /v1/vendor-user/update:
    put:
      consumes:
      - application/json
      description: API that update vendor-user
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateVendorUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UpdateVendorUserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update vendor-user
      tags:
      - vendor-user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"